refactor: 인수테스트 수정(Factory 클래스 삭제)
