+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -4,24 +4,21 @@ import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.line.dto.SectionRequest;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.stream.Collectors;
 
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_ID_추출;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_IDs_추출;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_목록_IDs_추출;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_목록_응답됨;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_목록_포함됨;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_삭제됨;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_생성_실패됨;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선_생성됨;
-import static nextstep.subway.line.TestLineAcceptanceFactory.지하철_노선과_종점역정보_파라미터_생성;
-import static nextstep.subway.station.TestStationAcceptanceFactory.지하철_역_생성;
-import static nextstep.subway.station.TestStationAcceptanceFactory.지하철_역_파라미터_생성;
+import static nextstep.subway.station.StationAcceptanceTest.지하철_역_생성;
+import static nextstep.subway.station.StationAcceptanceTest.지하철_역_파라미터_생성;
+import static org.assertj.core.api.Assertions.assertThat;
 
 @DisplayName("지하철 노선 관련 기능")
 public class LineAcceptanceTest extends AcceptanceTest {
@@ -93,7 +90,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         long 신분당선_ID = 지하철_노선_ID_추출(지하철_노선_생성_요청_응답);
 
         // when
-        ExtractableResponse<Response> 지하철_노선_조회_요청_응답 = 조회_요청(LINE_ROOT_PATH + 신분당선_ID);
+        ExtractableResponse<Response> 지하철_노선_조회_요청_응답 = 조회_요청(LINE_ROOT_PATH + "/" + 신분당선_ID);
 
         // then
         지하철_노선_목록_응답됨(지하철_노선_조회_요청_응답);
@@ -105,7 +102,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         long 신분당선_ID = 1L;
 
         // when
-        ExtractableResponse<Response> 지하철_노선_조회_요청_응답 = 조회_요청(LINE_ROOT_PATH + 신분당선_ID);
+        ExtractableResponse<Response> 지하철_노선_조회_요청_응답 = 조회_요청(LINE_ROOT_PATH + "/" + 신분당선_ID);
 
         // then
         지하철_노선_생성_실패됨(지하철_노선_조회_요청_응답);
@@ -121,7 +118,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         LineRequest 이호선_강남역_양재역 = 지하철_노선과_종점역정보_파라미터_생성("이호선", "green", 강남역.getId(), 양재역.getId(), 10);
 
         // when
-        ExtractableResponse<Response> 지하철_노선_수정_요청_응답 = 수정_요청(LINE_ROOT_PATH + 신분당선_ID, 이호선_강남역_양재역);
+        ExtractableResponse<Response> 지하철_노선_수정_요청_응답 = 수정_요청(LINE_ROOT_PATH + "/" + 신분당선_ID, 이호선_강남역_양재역);
 
         // then
         지하철_노선_목록_응답됨(지하철_노선_수정_요청_응답);
@@ -135,9 +132,54 @@ public class LineAcceptanceTest extends AcceptanceTest {
         long 신분당선_ID = 지하철_노선_ID_추출(지하철_노선_생성_요청_응답);
 
         // when
-        ExtractableResponse<Response> 지하철_노선_제거_요청_응답 = 삭제_요청(LINE_ROOT_PATH + 신분당선_ID);
+        ExtractableResponse<Response> 지하철_노선_제거_요청_응답 = 삭제_요청(LINE_ROOT_PATH + "/" + 신분당선_ID);
 
         // then
         지하철_노선_삭제됨(지하철_노선_제거_요청_응답);
+    }
+
+    public static List<Long> 지하철_노선_목록_IDs_추출(ExtractableResponse<Response> response) {
+        return response.jsonPath().getList(".", LineResponse.class).stream()
+                .map(LineResponse::getId)
+                .collect(Collectors.toList());
+    }
+
+    @SafeVarargs
+    public static List<Long> 지하철_노선_IDs_추출(ExtractableResponse<Response>... createResponse) {
+        return Arrays.stream(createResponse)
+                .map(it -> Long.parseLong(it.header("Location").split("/")[2]))
+                .collect(Collectors.toList());
+    }
+
+    public static long 지하철_노선_ID_추출(ExtractableResponse<Response> createResponse) {
+        return Long.parseLong(createResponse.header("Location").split("/")[2]);
+    }
+
+    public static void 지하철_노선_목록_포함됨(List<Long> expectedLineIds, List<Long> resultLineIds) {
+        assertThat(resultLineIds).containsAll(expectedLineIds);
+    }
+
+    public static LineRequest 지하철_노선과_종점역정보_파라미터_생성(String name, String color, Long upStationId, Long downStationId, int distance) {
+        return LineRequest.of(name, color, upStationId, downStationId, distance);
+    }
+
+    public static SectionRequest 종점역정보_파라미터_생성(Long upStationId, Long downStationId, int distance) {
+        return SectionRequest.of(upStationId, downStationId, distance);
+    }
+
+    public static void 지하철_노선_생성됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+    }
+
+    public static void 지하철_노선_생성_실패됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
+    }
+
+    public static void 지하철_노선_목록_응답됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+    }
+
+    public static void 지하철_노선_삭제됨(ExtractableResponse<Response> response) {
+        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
     }
 }
