--- src/test/java/nextstep/subway/line/TestLineAcceptanceFactory.java
@@ -1,70 +1,0 @@
-package nextstep.subway.line;
-
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import nextstep.subway.line.dto.LineRequest;
-import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.line.dto.SectionRequest;
-import org.springframework.http.HttpStatus;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-public class TestLineAcceptanceFactory {
-
-    public static List<Long> 지하철_노선_목록_IDs_추출(ExtractableResponse<Response> response) {
-        return response.jsonPath().getList(".", LineResponse.class).stream()
-                .map(LineResponse::getId)
-                .collect(Collectors.toList());
-    }
-
-    @SafeVarargs
-    public static List<Long> 지하철_노선_IDs_추출(ExtractableResponse<Response>... createResponse) {
-        return Arrays.stream(createResponse)
-                .map(it -> Long.parseLong(it.header("Location").split("/")[2]))
-                .collect(Collectors.toList());
-    }
-
-    public static long 지하철_노선_ID_추출(ExtractableResponse<Response> createResponse) {
-        return Long.parseLong(createResponse.header("Location").split("/")[2]);
-    }
-
-    public static void 지하철_노선_목록_포함됨(List<Long> expectedLineIds, List<Long> resultLineIds) {
-        assertThat(resultLineIds).containsAll(expectedLineIds);
-    }
-
-    public static LineRequest 지하철_노선과_종점역정보_파라미터_생성(String name, String color, Long upStationId, Long downStationId, int distance) {
-        return LineRequest.of(name, color, upStationId, downStationId, distance);
-    }
-
-    public static SectionRequest 종점역정보_파라미터_생성(Long upStationId, Long downStationId, int distance) {
-        return SectionRequest.of(upStationId, downStationId, distance);
-    }
-
-    public static void 지하철_노선_생성됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    public static void 지하철_노선_생성_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    public static void 지하철_노선_목록_응답됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 지하철_노선_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
-    }
-
-    public static void 지하철_구간_삭제됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
-    }
-
-    public static void 지하철_구간_삭제_실패됨(ExtractableResponse<Response> response) {
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-}
